requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pandas==2.3.1
numpy==2.0.0
azure-storage-blob==12.20.0
lightgbm==4.3.0
xgboost==2.1.1
scikit-learn==1.4.2
joblib==1.4.2
python-dotenv==1.0.1
requests==2.32.3
pyarrow>=15